Fix clang -Wtautological-constant-out-of-range-compare in SpinFrame (Android CI)
SpinFrame uses a uint16_t FrameType to bound the random search for a
QUIC_FRAME_IS_KNOWN value. That macro also compares against seera's
address-discovery frame types (QUIC_FRAME_ADD_ADDRESS_V4/REMOVE_ADDRESS,
> 0x3d0000), which don't fit in uint16_t. clang (Android/Darwin/macos) flags
those comparisons as tautological under -Werror; gcc (Linux x64) does not,
which is why local Linux builds passed.

Widen the value for the macro check ((uint64_t)FrameType). The random search
still uses uint16_t (so it terminates); frame types above the 16-bit range are
simply never generated.

Verified: clang++ -fsyntax-only with the CI Android flag set compiles
SpinFrame.cpp clean; BbrTest/CubicTest/SettingsTest/PathTest/quic_gtest also
clang-clean.

+++ src/core/unittest/SpinFrame.cpp
@@ -84,7 +84,13 @@ TEST(SpinFrame, SpinFrame1000000)
 
         do {
             TEST_QUIC_SUCCEEDED(CxPlatRandom(sizeof(FrameType), &FrameType));
-        } while (!QUIC_FRAME_IS_KNOWN(FrameType));
+            //
+            // Widen for the check so comparisons against frame types that don't
+            // fit in 16 bits (e.g. the address-discovery frames) aren't flagged
+            // as tautological by clang. Frame types above the 16-bit range are
+            // simply never generated here.
+            //
+        } while (!QUIC_FRAME_IS_KNOWN((uint64_t)FrameType));
 
         switch(FrameType) {
             case QUIC_FRAME_PADDING:
